#17 added login/logout
