+++ teque/config/settings.py
@@ -1,15 +1,3 @@
-"""
-Django settings for config project.
-
-Generated by 'django-admin startproject' using Django 3.1.7.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/3.1/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/3.1/ref/settings/
-"""
-
 from pathlib import Path
 import os
 
@@ -35,6 +23,7 @@ ALLOWED_HOSTS = ["*"]
 # Application definition
 
 INSTALLED_APPS = [
+    'testapp.apps.TestappConfig',
     'accounts.apps.AccountsConfig',
     'question.apps.QuestionConfig',
     'django.contrib.admin',
