+++ teque/question/views.py
@@ -1,7 +1,3 @@
 from django.shortcuts import render
-from django.contrib.auth.decorators import login_required
 
-
-@login_required  # 追加
-def post_list(request):
-    return render(request, 'question/post_list.html',  {})
+# Create your views here.
